+++ src/Caliburn.Light.Avalonia/Services/IWindowManager.cs
@@ -1,3 +1,5 @@
+using Avalonia.Platform.Storage;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace Caliburn.Light.Avalonia;
@@ -28,4 +30,28 @@ public interface IWindowManager
     /// <param name="viewModel">The view model of the window.</param>
     /// <returns>true if the window was successfully activated; otherwise, false.</returns>
     bool Activate(object viewModel);
+
+    /// <summary>
+    /// Displays an open file picker dialog.
+    /// </summary>
+    /// <param name="options">The dialog options.</param>
+    /// <param name="ownerViewModel">The owner view model.</param>
+    /// <returns>A list of selected files.</returns>
+    Task<IReadOnlyList<IStorageFile>> ShowOpenFilePickerAsync(FilePickerOpenOptions options, object ownerViewModel);
+
+    /// <summary>
+    /// Displays a save file picker dialog.
+    /// </summary>
+    /// <param name="options">The dialog options.</param>
+    /// <param name="ownerViewModel">The owner view model.</param>
+    /// <returns>The selected file</returns>
+    Task<IStorageFile?> ShowSaveFilePickerAsync(FilePickerSaveOptions options, object ownerViewModel);
+
+    /// <summary>
+    /// Displays an open folder picker dialog.
+    /// </summary>
+    /// <param name="options">The dialog options.</param>
+    /// <param name="ownerViewModel">The owner view model.</param>
+    /// <returns>A list of selected folders.</returns>
+    Task<IReadOnlyList<IStorageFolder>> ShowOpenFolderPickerAsync(FolderPickerOpenOptions options, object ownerViewModel);
 }